o Fixed bug in initialization of integer scores in a CM_TR_SCAN_MX, it   was using float transition scores instead of the integer ones. This   was not detected earlier because HMM banded matrices are used by   default, but now I'm testing cmsearch --max, which uses non-banded   matrices.
o Bugs still exist in Diana's TrCYK_DnC(), which I need for
  nonbanded alignment of truncated hits with --max. I'm worried these
  will be difficult to fix because I'm unfamiliar with that code. 

o Changed type of ExpInfo_t's <cur_eff_dbsize> from long to a
  double. This number is the effective number of sequences in the
  database, the E value of a hit is obtained by multiplying P and this
  number. I'm not sure why it was a long previously, but I guess it is
  because actual database size (# res) was stored as a long. But using
  a long was problematic if the database was smaller than the average
  size of a hit because <cur_eff_dbsize> would be defined as 0.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3891 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/cm_file.c
@@ -2475,7 +2475,7 @@ read_asc_1p0_cm(CM_FILE *cmfp, int read_fp7, ESL_ALPHABET **ret_abc, CM_t **opt_
 	  if (! is_real(tok))                                    ESL_XFAIL(eslEFORMAT, cmfp->errbuf, "tailp is not real number on E-xx line");
 	  cm->expA[exp_mode]->tailp = atof(tok);
 	  
-	  cm->expA[exp_mode]->cur_eff_dbsize = (long) (cm->expA[exp_mode]->nrandhits);
+	  cm->expA[exp_mode]->cur_eff_dbsize = (double) (cm->expA[exp_mode]->nrandhits);
 	  /* Previous line is to set cur_eff_dbsize as if database was of size cm->stats->expAA[p]->dbsize, we 
 	   * act as if the max hits we'll see is nrandhits, the number of hits we saw in cmcalibrate,
 	   * so this is the highest possible E-value we can get.
